third_party/WebKit/LayoutTests/fast/dom/elementsFromPoint/elementsFromPoint-iframes-expected.txt

PASS document.elementsFromPoint(1, 1).indexOf(iframe) is >= 0
PASS All elements in document.elementsFromPoint(1, 1) intersect (1, 1)
PASS document.elementsFromPoint(203, 1).indexOf(iframe) is >= 0
PASS All elements in document.elementsFromPoint(203, 1) intersect (203, 1)
PASS document.elementsFromPoint(1, 203).indexOf(iframe) is >= 0
PASS All elements in document.elementsFromPoint(1, 203) intersect (1, 203)
PASS document.elementsFromPoint(203, 203).indexOf(iframe) is >= 0
PASS All elements in document.elementsFromPoint(203, 203) intersect (203, 203)
PASS document.getElementById('iframe').contentDocument.elementsFromPoint(9, 9).indexOf(document.getElementById('iframe').contentDocument.getElementById('div')) is >= 0
PASS All elements in document.getElementById('iframe').contentDocument.elementsFromPoint(9, 9) intersect (9, 9)
PASS document.getElementById('iframe').contentDocument.elementsFromPoint(107, 9).indexOf(document.getElementById('iframe').contentDocument.getElementById('div')) is >= 0
PASS All elements in document.getElementById('iframe').contentDocument.elementsFromPoint(107, 9) intersect (107, 9)
PASS document.getElementById('iframe').contentDocument.elementsFromPoint(9, 107).indexOf(document.getElementById('iframe').contentDocument.getElementById('div')) is >= 0
PASS All elements in document.getElementById('iframe').contentDocument.elementsFromPoint(9, 107) intersect (9, 107)
PASS document.getElementById('iframe').contentDocument.elementsFromPoint(107, 107).indexOf(document.getElementById('iframe').contentDocument.getElementById('div')) is >= 0
PASS All elements in document.getElementById('iframe').contentDocument.elementsFromPoint(107, 107) intersect (107, 107)
PASS document.elementsFromPoint(51, 51).indexOf(iframeScrollbars) is >= 0
PASS All elements in document.elementsFromPoint(51, 51) intersect (51, 51)
PASS document.elementsFromPoint(203, 51).indexOf(iframeScrollbars) is >= 0
PASS All elements in document.elementsFromPoint(203, 51) intersect (203, 51)
PASS document.elementsFromPoint(51, 203).indexOf(iframeScrollbars) is >= 0
PASS All elements in document.elementsFromPoint(51, 203) intersect (51, 203)
PASS document.elementsFromPoint(203, 203).indexOf(iframeScrollbars) is >= 0
PASS All elements in document.elementsFromPoint(203, 203) intersect (203, 203)
PASS document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(1, 1).indexOf(document.getElementById('iframeScrollbars').contentDocument.getElementById('div')) is >= 0
PASS All elements in document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(1, 1) intersect (1, 1)
PASS document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(99, 1).indexOf(document.getElementById('iframeScrollbars').contentDocument.getElementById('div')) is >= 0
PASS All elements in document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(99, 1) intersect (99, 1)
PASS document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(1, 99).indexOf(document.getElementById('iframeScrollbars').contentDocument.getElementById('div')) is >= 0
PASS All elements in document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(1, 99) intersect (1, 99)
PASS document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(99, 99).indexOf(document.getElementById('iframeScrollbars').contentDocument.getElementById('div')) is >= 0
PASS All elements in document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(99, 99) intersect (99, 99)
PASS document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(-1, -1).length is 0
PASS document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(1, 1).length is 2
PASS document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(500, 500).length is 0
PASS document.getElementById('iframeScrollbars').contentDocument.elementsFromPoint(140, 140).length is 0
PASS successfullyParsed is true

TEST COMPLETE
